+++ src/observable.rs
@@ -6,6 +6,7 @@ pub(crate) mod from_future;
 pub use from_future::{from_future, from_future_with_err};
 pub(crate) mod interval;
 pub use interval::{interval, interval_at};
+pub(crate) mod connectable_observable;
 
 /// A representation of any set of values over any amount of time. This is the
 /// most basic building block rxrust